backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==21.1
Pillow==9.5.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
requests==2.31.0 